Add additional component input to form-input component to accept only Integer values
Description
==============
HTML input from type`number` by default accepts floating number as an user input. There wasn't a way to accept only Integer values in our vcd-form-input for now. To cover this case, a new component input `allowOnlyIntegers` is added, based on which a validation if the user input is an Ineteger is made. If the input is not an Integer an error `Inavid number` is shown. By default the component input's value is false.

Examples are updated.